// _archive/jellysea/src/utils/types.ts
export enum MediaStatus {
  UNKNOWN = 1,
  PENDING = 2,
  PROCESSING = 3,
  PARTIALLY_AVAILABLE = 4,
  AVAILABLE = 5,
  DELETED = 6,
}

export enum MediaType {
  MOVIE = 'movie',
  TV = 'tv',
}

export enum MediaRequestStatus {
  PENDING = 1,
  APPROVED = 2,
  DECLINED = 3,
}

export interface DownloadingItem {
  externalId: number
  downloadId: string
  size: number
  sizeLeft: number
  status: string
  timeLeft: string
  estimatedCompletionTime: string
  episode?: {
    seasonNumber: number
    episodeNumber: number
  }
}

export interface Season {
  id: number
  seasonNumber: number
  status: MediaStatus
}

export interface User {
  id: number;
  email: string;
  username?: string;
  plexUsername?: string;
  jellyfinUsername?: string;
  displayName: string;
  avatar: string;
  permissions: number;
  userType: number;
  requestCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface MediaInfo {
  id: number;
  tmdbId: number;
  tvdbId?: number;
  mediaType: 'movie' | 'tv';
  status: MediaStatus;
  status4k?: MediaStatus;
  requests: MediaRequest[];
  plexUrl?: string;
  plexUrl4k?: string;
  iOSPlexUrl?: string;
  iOSPlexUrl4k?: string;
  serviceUrl?: string;
  serviceUrl4k?: string;
  downloadStatus?: DownloadingItem[];
  downloadStatus4k?: DownloadingItem[];
  mediaAdded?: string;
  createdAt: string;
  updatedAt: string;
  seasons?: Season[];
}

export interface MediaRequest {
  id: number;
  status: number;
  media: MediaInfo;
  type?: 'movie' | 'tv';
  seasons?: Season[];
  createdAt: string;
  updatedAt: string;
  requestedBy: User;
  is4k: boolean;
}

export interface MovieResult {
  id: number;
  mediaType: 'movie';
  popularity: number;
  posterPath?: string;
  backdropPath?: string;
  voteCount: number;
  voteAverage: number;
  genreIds: number[];
  overview: string;
  originalLanguage: string;
  title: string;
  originalTitle: string;
  releaseDate: string;
  adult: boolean;
  mediaInfo?: MediaInfo;
}

export interface TvResult {
  id: number;
  mediaType: 'tv';
  popularity: number;
  posterPath?: string;
  backdropPath?: string;
  voteCount: number;
  voteAverage: number;
  genreIds: number[];
  overview: string;
  originalLanguage: string;
  name: string;
  originalName: string;
  originCountry: string[];
  firstAirDate: string;
  mediaInfo?: MediaInfo;
}